Write tests for Candle record syntax
Record syntax will be implemented as syntactic sugar in the parser.
Declaring a record is done like this:

  type my_rec = Constr of { f1: t1; f2: t2; ...; fN: tN };;

The above declaration is parsed into a datatype declaration and a series
of function declarations for projection, update and construction:

  type my_rec = Constr t1 t2 ... tN
  (* Here, we assume that f1 < f2 < ... < fN; otherwise field names and
     types are sorted and re-ordered *)

  let __f1_record_upd (Constr f1 f2 ... fN) f1 = Constr f1 f2 ... fN;;
  let __f2_record_upd (Constr f1 f2 ... fN) f2 = ...;;
  ...
  let __f1_record_proj (Constr f1 f2 ... fN) = f1;;
  let __f2_record_proj (Constr f1 f2 ... fN) = f2;;
  ...
  let __record_constr_Constr_f1_f2_..._fN =
    fun f1 -> fun f2 -> ... -> fun fN -> Constr f1 f2 ... fN;;

Records with non-distinct field names are rejected. The fields and types
are ordered by the names of the fields. Internally, all records are
regular datatypes. This makes it possible (but unreliable) to pattern
match. For example:

  type myrec = Foo { b: int; a: bool };;

  let foo (Foo a b) = ... (* correct *) ;;
  let foo' (Foo b a) = ... (* incorrect *) ;;

Projection is done with a dot:

  myRecValue . foo (* becomes __foo_record_proj myRecValue *)

'with' updates:

  myRecValue with foo := bar
  (* becomes __foo_record_upd myRecValue bar *)