Create record funs in TypeDefinition function
Changed the function that deals with type definitions to also generate a
sequence of projection/update/constructor functions for record type
constructors. The functions are given names that are readable (in case
they pop up in an error message) but impossible to get past the parser.

Some syntactic criteria are checked (e.g., all field names under a
single record constructor must be distinct), but other than that, the
standard semantics is to simply overshadow previous definitions with the
same name.