Prove some syntax properties
inst_ok_less requires that some of the registers used with
the LongMul instruction are distinct, depending on platform.
Because parts of the Pancake compiler populate omitted arguments
with defaults, this property is not guaranteed to hold unless
the source program has no such arity mismatches.
This assumption is propagated all the way to the top.

This is not really a limitation: offending programs are
impossible to even write in concrete syntax.