Improve flat_pattern compiler
This commit addresses an issue where the flat_pattern compiler (that
removes pattern matches) would create code that would confuse the
analysis of the clos_known pass. These things are changed:
* Guard simplification is done bottom-up (instead of top-down)
* The compiler won't insert unncessary conditionals (i.e., if true then
  ... or if false then ...).

This commit leaves a few cheats in flat_patternProof related to
set_globals and esgc_free. These theorems are currently equality
theorems, but the new compiler can shrink these sets if it removes a
branch in a conditional; they should be turned into subset theorems.