Add parser support for let recs/ands w/o variables
This adds parser support for these sort of declarations:

  let rec f (* no variables *) = ...

and

  let rec f ... = ...
  and g (* no variables *) = ...

by assuming that these are always functions and eta expanding them.
This enables one to write these sort of functions:

  let rec f =
    function
    | Foo -> ...
    | ... -> ...

omitting the variable. Unfortunately, it also makes 'g' a function in
this declaration:

  let rec f x = x + 1
  and g = 3

This is different from OCaml which can deal with let rec chains
containing values.