Change tail call syntax
Make tail calls be of the form

  return f();

...where they used to be of the form

  f();

Previously, if one wrote something like this:

  fun main() {
    f();
    f();
  }

  fun f() { ... }

...then the second call to f will never be executed,
because the tail call eats up the stack frame and makes main return
whatever the first call returned. This is confusing.

Stand-alone function calls are still allowed, but parse to
returning calls that save the return value in the variable "";
this should always be fresh since nothing parses to an empty variable.