Draft compilation of declaring calls
I first expected to need shape annotations on functions, but
this is a dead end: function pointers make it so that we cannot
statically determine which function is being called.

Instead I've put shape annotations on the declaring calls themselves.
At some point in the future we should probably make declarations
available (mandatory or optional) on all declarations.