pancake: patch crep_to_loop to use static Call
Adjust the crep_to_loop compilation phase to spot the case where the target is
a constant "Label X" expression, and produce a call to a known target on the
loopLang side. Hopefully this should result in static calls in the final
binary, rather than needless indirect calls. Frustratingly, for a simple
change, quite a lot of the proof script had to be amended.