Broaden simp_if pass of word_simp
The word_simp pass already contains a simp_if component that looks for
adjacent If constructs which really have two control-flow paths, not four.

For instance, rewrite
  (If C X Y; If C2 X2 Y2) -> (If C (X ; X2) (Y ; Y2))

This patch reuses other logic present in word_simp to make this conversion much
more general. It permits a small number of intermediate simple ops.

For instance, rewrite
  (If C X Y; Zs; If C2 X2 Y2) -> (If C (X ; Zs; X2) (Y ; Zs; Y2))

This version also confirms that X; Zs -> (C2 == True) using the word_simp
const-propagation pass rather than recognizing a simple syntactic form.

This should work more generally and reduce the number of branches in compiled
pancake programs. The proof is not complicated, however, the functional program
that recognizes the case is.