Extend the double FFI
This commit extends the double ffi with new functions (conversions from
and to 64-bit signed integers, exp, ln, floor and pow). It adds error
detection to fromString (in case the floating point parse failed), and
changes the type to Double.fromString : string -> double option.

It does *not* add the new functions to the DoubleFFI FFI model, so they
may only by usable from the REPL.