Remove float literals from Candle
The parser previously called Double.fromString on strings to produce
floating point literals. This conversion happened at run-time, and the
compiler was unable to optimize it away, resulting in *expensive* calls
to sprintf at every use of a float literal. Until there is a real
floating point parser in CakeML, the literals are disabled in the
parser.