Remove DoubleProofScript.sml
This theory contains some proofs about the Double FFI which are no
longer true. The theorems basically state that the fromString and
toString functions are inverses, which isn't exactly true, either
(they can fail in various ways).

Nothing in the codebase makes use of this theory, so it should be safe
to delete it.