Rework printing and clean up tests
Initially we assumed that e in (Print e) was always a string. However,
according to the Dafny reference this is not a correct assumption. This
commit adds support to print ints, bools, strings and sequences. Note
that we print in a format that is equivalent to the output of
`dafny run file.dfy` to make testing easier.