Fix some word syntactic properties in pancake
Ugh, of course the pancake proofs basically replay all of the word syntactic
stuff yet again, making even heavier work of it that word_to_stack or
backendProof. I've fixed some of it, tweaked some really slow proofs, and
commented out some proofs that don't seem to be used anywhere. There are still
too many slow proofs to test all of this on my laptop - committing to move it
to a server-class machine.