Always encode Dafny strings as CakeML char lists
In Dafny, `string` and `seq<char>` are synonyms. In CakeML, `string`
and `char list` are not, which lead to type errors. With this commit,
we encode both types as `char list`.

Since our implementation keeps track of Dafny types in type
environments, we also rewrite `Primitive String` to `Seq (Primitive
Char)`. Forgetting to do so using `replace_string_type` has the
potential to become a popular source of bugs in the future :P