Add newtype support by interpreting them as their base type
One concern I had was whether we would run into issues due to the fact
that we store "normalized types" (i.e. base types for newtypes) in the
type environment, for example when doing a type test. However, it
appears that such things are not legal:

```dfy
newtype i32 = x: int | -0x8000_0000 <= x < 0x8000_0000

method Main() {
    var i:i32 := 1;
    print (i is int);
}
```

does not run - instead, it returns the error
```
test.dfy(5,10): Error: type test for type 'int' must be from an expression assignable to it (got 'i32')
  |
5 | 	print (i is int);
  |           ^^
```
which matches my interpretation of the reference.