Progress up to TailCall
The generalised return values clarify what's going on a bit, and the Hoare
logic is now sound and general for various syntactic operators. Function calls
are covered, but only in the TailCall sense. The big obstacle is where to
shelve the existing locals when performing a normal call. There were thoughts
about doing this via local renaming, but it's all a bit too ugly.