Init-array example and Hoare verification
Add an example program that does a trivial initialisation of an array. This is
verified in "classic" evaluate style (with a helper lemma doing the induction
over the loop) and also in Hoare style. A tactic is scratched together on the
spot to drive the Hoare logic. The tactic is a extremely slow and everything is
a bit wordy, but it is workable.