Tweak Pancake annotations
The main change is to adjust the type of the Annot constructor to take two
mlstring arguments rather than one. The annotations are already being used for
two different roles, and I've just thought of a third. The distinction allows
the first string to identify the kind of annotation, which lets annotation
consumers tell them apart without doing fiddly parsing.

Also, the explorer output prints the strings as strings (with quotes and
escapes for some characters) rather than including them verbatim, so the output
s-expression style paren-expression is guaranteed to be well-formed.