Cleanup and tweaks, explicit reverse order
Some simplifications and tweaks, and the performance is roughly acceptable
now.

The big change is to introduce a rev_hoare constant that makes explicit that
the goal is to discover a precondition (to the "second thing") and then prove
some goal continuation. With this, the rules can naturally be applied using
match_mp_tac rather than with irule_at, which seems simpler and better.