Hoare-pancake: cleanup
Commit to the rev_hoare style, and add some overcomplicated code to the
panHoareScript file to generate rev_hoare rules from the inductive rules (to
avoid having two copies of every rule). Clean up a lot of leftover attempts
which remain in the git history in case they're ever of interest.