[Dafny] Fix compilation of return
In the case of `Return e`, we now write `e` to a reference, raise
an exception, and then "return" `e` from the function by
dereferencing it. This is (probably) more correct, and allows for
some simplification in `from_stmts`, where we do not need to treat
singleton lists differently anymore.