[Dafny] Fix MetCall semantics + freshen proof
We were always restoring the locals, even when the call succeeded. Now, we only
restore locals when failing to read them, or when the amount of lhs and rhs does
not match up.

Also, seems like the freshen proof was broken from our last change to the While
semantics. That's fixed, in addition to what broke due to the MetCall change.