Replace export_rewrites with simp where possible (#1264)
There are a few things worth pointing out:

1. I've slightly restructured some new_specification calls. In particular, I
factored out the proof into a local theorem, and more importantly, I factored
out the string name into a variable (see clFFIScript.sml). The motivation for
the latter was to ultimately ban `export_rewrites["`, but it turns out that
there are still good uses for this (for example when adding definitions/theorems
from other theories to the simpset). Arguably, it gets rid of some duplication,
which is why I haven't reverted it.

2. I've adjusted various tutorial files to use simp tags, since presumably we
want newcomers to be exposed to that. simple_bstScript.sml still mentions what
it desugars to at some point, so it should be fine.

3. I deleted astPropsScript.sml -- there was a TODO that said to do it, and it
also had commented out export_rewrites.